refactor(audio): streamline voice tag lifecycle, storage cleanup, and TTS adapter

+++ Backend/modules/audio/src/main/java/com/pwb/audio/infrastructure/tts/properties/GoogleTtsProperties.java
@@ -15,7 +15,7 @@ public class GoogleTtsProperties {
 
     private String audioEncoding = "MP3";
 
-    private Integer speakingRate = 1;
+    private Float speakingRate = 1.0f;
 
-    private Double pitch = 0.0;
+    private Float pitch = 0.0f;
 }
